Install nvim-cmp
This change adds nvim-cmp with multiple sources. Namely, this
significantly augments the capabilities of the LSP. I can now view the
docs, complete functions with argument placeholders and get access to
overloads.

The snippet engine seems to not work well though. There are some
snippets, but seem to come from the LSP.